+++ src/main/java/insane96mcp/survivalreimagined/mixin/BaseSpawnerMixin.java
@@ -1,6 +1,6 @@
 package insane96mcp.survivalreimagined.mixin;
 
-import insane96mcp.survivalreimagined.module.misc.feature.TempSpawner;
+import insane96mcp.survivalreimagined.module.misc.feature.Spawners;
 import net.minecraft.core.BlockPos;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.world.level.BaseSpawner;
@@ -15,12 +15,12 @@ public class BaseSpawnerMixin {
 
 	@Inject(at = @At("HEAD"), method = "serverTick")
 	private void serverTick(ServerLevel p_151312_, BlockPos p_151313_, CallbackInfo callback) {
-		TempSpawner.onServerTick((BaseSpawner) (Object) this);
+		Spawners.onSpawnerServerTick((BaseSpawner) (Object) this);
 	}
 
 	@Inject(at = @At("HEAD"), method = "clientTick")
 	private void serverTick(Level p_151320_, BlockPos p_151321_, CallbackInfo callback) {
-		TempSpawner.onClientTick((BaseSpawner) (Object) this);
+		Spawners.onClientTick((BaseSpawner) (Object) this);
 	}
 
 }